backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9